requirements.txt
firebase_admin==6.5.0
fitz==0.0.1.dev2
numpy==1.26.4
pandas==2.2.2
protobuf==3.20.3
PyPDF2==3.0.1
Pyrebase4==4.8.0
streamlit==1.32.0
tqdm==4.66.2
transformers==4.39.2
protobuf==3.20.3
PyMuPDF==1.24.3
